+++ common/src/main/java/net/lizistired/animationoverhaul/mixin/config/MixinPlayerSkinProvider.java
@@ -1,12 +1,13 @@
-package gg.moonflower.animationoverhaul.mixin.config;
+package net.lizistired.animationoverhaul.mixin.config;
 
 import java.io.File;
-import net.minecraft.client.resources.SkinManager;
+
+import net.minecraft.client.texture.PlayerSkinProvider;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.gen.Accessor;
 
-@Mixin({SkinManager.class})
+@Mixin({PlayerSkinProvider.class})
 public interface MixinPlayerSkinProvider {
-    @Accessor("skinsDirectory")
+    @Accessor("skinCacheDir")
     File getSkinCacheDirectory();
 }
