+++ common/src/main/java/net/lizistired/animationoverhaul/mixin/debug/MixinAbstractClientPlayer.java
@@ -17,6 +17,8 @@ public class MixinAbstractClientPlayer {
 
     @Inject(method = "getCapeTexture", at = @At("HEAD"), cancellable = true)
     private void useDebugCapeTexture(CallbackInfoReturnable<Identifier> cir) {
+        if (AnimationOverhaulMain.getConfig().isEnableDebugCape()) {
             cir.setReturnValue(debugCapeLocation);
+        }
     }
 }
